refactor create order page
