+++ src/components/pages/order/index.ts
@@ -6,11 +6,12 @@ export * from './menu-dropdown';
 export * from './offered-by-formula';
 export * from './options-dropdown';
 export * from './shared-link';
-export * from './shop-owner';
+export * from './shop-owner-image';
+export * from './shop-owner-profile';
+export * from './shop-owner-transfer-info';
 export * from './show-formula';
 export * from './table';
 export * from './table-add-row-button';
 export * from './table-header';
 export * from './table-row';
 export * from './transfer-formula';
-export * from './transfer-info';
